requirements.txt
pandas
requests
pyyaml
aiohttp
aiofiles
asyncio